feat: Complete Cognitive Integration & De-Mocking
- Remove legacy CodeAtom schema (now using Atom with Modality='code')
- Implement RoslynAtomizer for semantic C# AST parsing
- Implement TreeSitterAtomizer for polyglot code parsing (Python, JS, Go, etc.)
- Wire ProvenanceController to real Neo4j services (remove mock data)
- Add FeedbackController for RLHF weight adjustment
- Create ContextRetrievalService for prompt hydration with spatial KNN
- Create Neo4jProvenanceQueryService with Cypher queries
- Fix ErrorStatistics types (long TotalErrors, double AverageClusterSize)
- Parameterize all deployment scripts (remove hardcoded values)
- Add environment variable support: HARTONOMOUS_SQL_SERVER, HARTONOMOUS_SQL_DATABASE
- Verify script idempotency (Deploy-All.ps1 runs twice without errors)
- All OODA Loop SPs verified: sp_Analyze, sp_Act, sp_Hypothesize, sp_Learn

+++ src/Hartonomous.Api/DTOs/MLOps/ErrorStatistics.cs
@@ -2,9 +2,9 @@ namespace Hartonomous.Api.DTOs.MLOps;
 
 public record ErrorStatistics
 {
-    public int TotalErrors { get; init; }
+    public long TotalErrors { get; init; }
     public int ClustersFound { get; init; }
-    public int AverageClusterSize { get; init; }
+    public double AverageClusterSize { get; init; }
     public required string SpatialSpread { get; init; }
     public required string TemporalWindow { get; init; }
     public required string MostCommonErrorType { get; init; }
